add divide matrix method

+++ 1_basic_method.cpp
@@ -13,7 +13,7 @@ vector< vector<int> > basic_method(const vector< vector<int> >& matrixA,
     vector< vector<int> > matrixC(M, vector<int>(K));
 	for (int i = 0; i < M; i++) {
 		for (int j = 0; j < N; ++j) {
-			double result = 0;
+			int result = 0;
 			for (int k = 0; k < K; ++k) {
 				result = result + matrixA[i][k] * matrixB[k][j];
 			}
